integrasi bootstrap, remove class permission, perbaikan handle NPE

+++ src/main/java/id/muhamadridwan/starter/services/impl/RoleServiceImpl.java
@@ -5,11 +5,11 @@
 package id.muhamadridwan.starter.services.impl;
 
 import id.muhamadridwan.starter.services.RoleService;
-import id.muhamadridwan.starter.models.Permission;
 import id.muhamadridwan.starter.models.Role;
-import id.muhamadridwan.starter.repositories.PermissionRepository;
 import id.muhamadridwan.starter.repositories.RoleRepository;
+import id.muhamadridwan.starter.utils.Util;
 import java.util.List;
+import javax.persistence.EntityNotFoundException;
 import javax.persistence.NoResultException;
 import javax.transaction.Transactional;
 import org.slf4j.Logger;
@@ -27,7 +27,6 @@ public class RoleServiceImpl implements RoleService {
 
     private static final Logger LOG = LoggerFactory.getLogger(RoleServiceImpl.class);
 
-    private PermissionRepository permissionRepository;
     private RoleRepository roleRepository;
 
     @Autowired
@@ -35,65 +34,55 @@ public class RoleServiceImpl implements RoleService {
         this.roleRepository = roleRepository;
     }
 
-    @Autowired
-    public void setPermissionRepository(PermissionRepository permissionRepository) {
-        this.permissionRepository = permissionRepository;
+    @Override
+    public void addRole(Role role) throws NullPointerException {
+        if (Util.isNotNull(role)) {
+            LOG.info("Add role : " + role.getName());
+            roleRepository.save(role);
+        }
     }
 
     @Override
-    public void addRole(Role role) {
-        LOG.info("Add role : " + role.getName());
-        roleRepository.save(role);
+    public void updateRole(Role role) throws NullPointerException, EntityNotFoundException {
+        if (Util.isNotNull(role) || Util.isNotNull(role.getId())) {
+            LOG.info("Update role : " + role.getName());
+            Role updateRole = roleRepository.findOne(role.getId());
+            updateRole.setName(role.getName());
+            roleRepository.save(updateRole);
+        }
     }
 
     @Override
-    public void updateRole(Role role) {
-        LOG.info("Update role : " + role.getName());
-        Role updateRole = roleRepository.findOne(role.getId());
-        updateRole.setName(role.getName());
-        roleRepository.save(updateRole);
+    public void deleteRole(Long id) throws NullPointerException, EntityNotFoundException {
+        if (Util.isNotNull(id)) {
+            Role deleteRole = roleRepository.getOne(id);
+            LOG.info("Delete role : " + deleteRole.getName());
+            roleRepository.delete(deleteRole);
+        }
     }
 
     @Override
-    public void deleteRole(Long id) {
-        Role deleteRole = getRole(id);
-        LOG.info("Delete role : " + deleteRole.getName());
-        roleRepository.delete(deleteRole);
-    }
-
-    @Override
-    public Role getRole(Long id) {
+    public Role getRole(Long id) throws NullPointerException, EntityNotFoundException {
         LOG.info("Get role by id : " + id);
-        return roleRepository.findOne(id);
+        Role role = null;
+        if (Util.isNotNull(role)) {
+            role = roleRepository.findOne(id);            
+        }
+        return role;
     }
 
     @Override
-    public Role getRole(String roleName) {
+    public Role getRole(String roleName) throws NullPointerException, EntityNotFoundException {        
         LOG.info("Get role by name : " + roleName);
-        return roleRepository.findByName(roleName);
+        Role role = null;
+        if (Util.isNotNull(roleName)) {
+            role = roleRepository.findByName(roleName);
+        }        
+        return role;       
     }
 
     @Override
     public List<Role> getRoles() {
         return roleRepository.findAll();
     }
-
-    @Override
-    public void assignPermissionToRole(Permission permission, Role role) {
-        LOG.info("Assign permission = { " + permission.getName() + " } to role -> " + role.getName());
-
-        Permission assignPermission = permissionRepository.findByName(permission.getName());
-        if (assignPermission == null) {
-            throw new NoResultException("Permission does not exist");
-        }
-
-        Role assignRole = roleRepository.findByName(role.getName());
-        if (assignRole == null) {
-            throw new NoResultException("Role does not exist");
-        }
-
-        assignRole.getPermission().add(assignPermission);
-        roleRepository.save(assignRole);
-    }
-
 }
